+++ CHANGELOG.md
@@ -9,6 +9,12 @@
 	* Removing thorns
 	* Opening Cursed Luggage
 	* Passing and Catching thrown items
+* Added penalties for:
+	* Slipping on a jellyfish
+	* Getting stung by a scorpion
+	* Getting hit by a tumbleweed
+* Added modifiers for:
+	* Being in the rain
 * Added a new config for Style Rank Multiplier, to change how many points are needed for each rank. Increase to make it take longer to rank up.
 * Added an internal trick cooldown to ensure that the same trick cannot be activated more than once every .1 seconds. Safety feature to avoid being spammed with tricks in glitched situations.